embedding/pyproject.toml
[tool.poetry]
name = "embedding"
version = "0.1.0"
description = ""
authors = ["Jonny Kernes <jmkernes@g.ucla.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.56.0"
grpcio-status = "^1.51.1"
grpcio-reflection = "^1.51.1"
grpc-interceptor = "^0.15.0"
opentelemetry-api = "^1.15.0"
opentelemetry-exporter-otlp = "^1.15.0"
opentelemetry-instrumentation-grpc = "^0.36b0"
optimum = "^1.9.1"
onnxruntime = "^1.15.1"
loguru = "^0.7.0"
transformers = "^4.31.0"
pydantic = "^2.0.3"
fire = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
